ResSimAD.Models.get_model — Function

get_model(example_name::String)

Get example model example_name.

Examples

julia> using ResSimAD: get_model

julia> sim, options = get_model("example1");

ResSimAD.SimMaster.runsim — Function

runsim(sim::Sim)

Run simulation from day sim.scheduler.t_current to day sim.scheduler.time_step[end]

Examples

julia> using ResSimAD: get_model, runsim

julia> sim, options = get_model("example1");

julia> sim.scheduler.t_current
0.0

julia> sim.scheduler.time_step[end]
1825.0

julia> runsim(sim);

julia> sim.scheduler.t_current
1825.0

ResSimAD.SimMaster.step_to — Function

step_to(sim::Sim, t::Float64)

Simluate until day t.

Examples

julia> using ResSimAD: get_model, step_to

julia> sim, options = get_model("example1");

julia> sim.scheduler.t_current
0.0

julia> step_to(sim, 100.0);

julia> sim.scheduler.t_current
100.0

ResSimAD.SimMaster.time_step — Function

time_step(sim::Sim)

Simluate for one time step. Time step length is sim.scheduler.dt.

Examples

julia> using ResSimAD: get_model, time_step

julia> sim, options = get_model("example1");

julia> sim.scheduler.t_current
0.0

julia> sim.scheduler.dt
0.1

julia> time_step(sim);

julia> sim.scheduler.t_current
0.1

ResSimAD.SimMaster.newton_step — Function

newton_step(sim::Sim)

Simulate for one newton step.

Examples

julia> using ResSimAD: get_model, newton_step, get_residual_error

julia> using Printf: @printf

julia> sim, options = get_model("example1");

julia> @printf("%.3e",get_residual_error(sim))
2.210e-01
julia> newton_step(sim)

julia> @printf("%.3e",get_residual_error(sim))
1.927e-02
julia> newton_step(sim)

julia> @printf("%.3e",get_residual_error(sim))
9.871e-05

ResSimAD.SimMaster.change_well_mode — Function

change_well_mode(sim::Sim, name::String, mode::String, target::Float64)

Change the control mode of well name to be mode with target value target

Arguments

sim::Sim: Sim object

name::String: name of the well

mode::String: target mode, not case sensitive

producers:

"bhp": constant BHP

"shut": shut-in

"orat": constant oil rate

"wrat": constant water rate

"lrat": constatnt liquid rate

injectors:

"bhp": constant BHP

"wrat": constant water rate

target::Float64

Examples

julia> using ResSimAD: get_model, change_well_mode, silence

julia> silence();

julia> sim, options = get_model("example1");

julia> change_well_mode(sim, "P1", "bhp", 6200.);

julia> println(sim.facility["P1"].mode)
CBHP

julia> println(sim.facility["P1"].target)
6200.0

ResSimAD.SimMaster.change_well_target — Function

change_well_target(sim::Sim, name::String, target::Float64)

Change the control target of well name to be target.

Arguments

sim::Sim: Sim object

name::String: name of the well

target::Float64: well control target

Examples

julia> using ResSimAD: get_model, change_well_target, silence

julia> silence();

julia> sim, options = get_model("example1");

julia> change_well_target(sim, "P1", 6200.);

julia> println(sim.facility["P1"].target)
6200.0

ResSimAD.SimMaster.get_well_rates — Function

get_well_rates(sim::Sim, name::String, data::String)

Get column data from dataframe sim.facility[name].results

Arguments

sim::Sim: Sim object

name::String: name of the well

data::String: column name (case insensitive)

"TIME": Time steps

"ORAT": Oil rate

"WRAT": Water rate

"GRAT": Gas rate

"LRAT": Liquid rate

"WBHP": BHP

Examples

julia> using ResSimAD: get_model, runsim, get_well_rates, silence

julia> silence();

julia> sim, options = get_model("example1");

julia> runsim(sim);

julia> tstep = get_well_rates(sim, "P1", "Time");

julia> qo = get_well_rates(sim, "P1", "ORAT");

julia> length(tstep) > 0
true
julia> length(qo) > 0
true

ResSimAD.SimMaster.get_data — Function

get_data(sim::Sim, dataname::String)

Get data for dataname dataname.

Examples

julia> using ResSimAD: get_model, step_to, get_data, silence

julia> silence();

julia> sim, options = get_model("example1");

julia> step_to(sim, 10.0);

julia> λo = get_data(sim, "λo");

julia> size(λo)
(450,)

ResSimAD.SimMaster.get_state_map — Function

get_state_map(sim::Sim, var::String, t::Float64)

Get state map for variable var at time t.

Examples

julia> using ResSimAD: get_model, runsim, get_state_map, get_well_rates, silence

julia> silence();

julia> sim, options = get_model("example1");

julia> runsim(sim);

julia> t = get_well_rates(sim, "P1", "TIME");

julia> po = get_state_map(sim, "po", t[end]);

julia> size(po)
(450,)

ResSimAD.SimMaster.add_well — Function

add_well(sim::Sim, welltype::String, well_option::Dict)

Add one well of type welltype, with well setups specified in well_option

Arguments

sim::Sim: Sim object

welltype::String: type of the well (case insensitive)

"injector": injector

"producer": producer

well_option::Dict: well setups with keys

"name" -> String, name of the well

"perforation" -> Vector{Tuple{Int, Int, Int}}, indices for perforated blocks

"radius" -> Float64, well radius

"mode" -> String, well control mode

"target" -> Float64, well constrol target

Examples

julia> using ResSimAD: get_model, add_well, silence

julia> silence();

julia> sim, options = get_model("example1");

julia> length(keys(sim.facility))
2

julia> p2 = Dict("name" => "P2", "perforation"=>[(5,5,1)], "radius"=>0.5, "mode"=>"bhp", "target"=>5600.);

julia> add_well(sim, "producer", p2);

julia> length(keys(sim.facility))
3

ResSimAD.SimMaster.shut_well — Function

shut_well(sim::Sim, name::String)

Shut well name

ResSimAD.SimMaster.change_dt — Function

change_dt(sim::Sim, dt::Float64)

Set next time step size to be dt

ResSimAD.silence — Function

silence()

Do not show any message except for "error" messages. See ResSimAD.verbose() for more information.

Examples

julia> using ResSimAD:silence

julia> silence();

ResSimAD.debug — Function

debug()

Show "debug" messages. See ResSimAD.verbose() for more information.

Examples

julia> using ResSimAD:debug

julia> debug();

ResSimAD.verbose — Function

verbose(level::String)

Set verbose level. Show messages with higher level than the set level. See Memento.jl for more information.

Arguments

level::String: verbose levels (low to high)

"debug": verbose message used for debugging

"info": general information

"notice": important events that are part of normal execution

"warn": warning that may cause the program to fail

"error": error that causes the program to terminate

Examples

julia> using ResSimAD:verbose

julia> verbose("debug");