ResSimAD.SimMaster.runsim — Function

runsim(sim::Sim; verbose=BRIEF)

Run simulation from day sim.scheduler.t_current to day sim.scheduler.time_step[end]

Examples

julia> using ResSimAD: get_model, runsim, SILENT

julia> sim, options = get_model("example1");

julia> sim.scheduler.t_current
0.0

julia> sim.scheduler.time_step[end]
1825.0

julia> runsim(sim, verbose=SILENT);

julia> sim.scheduler.t_current
1825.0

ResSimAD.SimMaster.step_to — Function

step_to(sim::Sim, t::Float64; verbose=BRIEF)

Simluate until day t.

Examples

julia> using ResSimAD: get_model, step_to, SILENT

julia> sim, options = get_model("example1");

julia> sim.scheduler.t_current
0.0

julia> step_to(sim, 100.0, verbose=SILENT);

julia> sim.scheduler.t_current
100.0

ResSimAD.SimMaster.time_step — Function

time_step(sim::Sim; verbose=BRIEF)

Simluate for one time step. Time step length is sim.scheduler.dt.

Examples

julia> using ResSimAD: get_model, time_step, SILENT

julia> sim, options = get_model("example1");

julia> sim.scheduler.t_current
0.0

julia> sim.scheduler.dt
0.1

julia> time_step(sim, verbose=SILENT);

julia> sim.scheduler.t_current
0.1

ResSimAD.SimMaster.newton_step — Function

newton_step(sim::Sim; verbose=BRIEF)

Simulate for one newton step.

Examples

julia> using ResSimAD: get_model, newton_step, get_residual_error, SILENT

julia> using Printf: @printf

julia> sim, options = get_model("example1");

julia> @printf("%.3e",get_residual_error(sim))
2.210e-01
julia> newton_step(sim)

julia> @printf("%.3e",get_residual_error(sim))
1.927e-02
julia> newton_step(sim)

julia> @printf("%.3e",get_residual_error(sim))
9.871e-05

ResSimAD.SimMaster.change_well_mode — Function

change_well_mode(sim::Sim, name::String, mode::String, target::Float64)

Change the control mode of well name to be mode with target value target

Arguments

sim::Sim: Sim object

name::String: name of the well

mode::String: target mode, not case sensitive

"bhp": constant BHP

"shut": shut-in

"orat": constant oil rate

"wrat": constant water rate

"lrat": constatnt liquid rate

target::Float64

Examples

julia> using ResSimAD: get_model, change_well_mode

julia> sim, options = get_model("example1");

julia> change_well_mode(sim, "P1", "bhp", 6200.);

julia> println(sim.facility["P1"].mode)
CBHP

julia> println(sim.facility["P1"].target)
6200.0

ResSimAD.SimMaster.change_well_target — Function

change_well_target(sim::Sim, name::String, target::Float64; cut_dt=false)

Change the control target of well name to be target. cut time step to be sim.scheduler.dt0 if cut_dt=true

Arguments

sim::Sim: Sim object

name::String: name of the well

target::Float64

cut_dt::Boolean

Examples

julia> using ResSimAD: get_model, change_well_target

julia> sim, options = get_model("example1");

julia> change_well_target(sim, "P1", 6200.);

julia> println(sim.facility["P1"].target)
6200.0

ResSimAD.SimMaster.get_well_rates — Function

get_well_rates(sim::Sim, name::String, data::String)

Get column data from dataframe sim.facility[name].results

Arguments

sim::Sim: Sim object

name::String: name of the well

data::String: column name

"Time": Time steps

"ORAT": Oil rate

"WRAT": Water rate

"GRAT": Gas rate

"LRAT": Liquid rate

"WBHP": BHP

Examples

julia> using ResSimAD: get_model, runsim, get_well_rates, SILENT

julia> sim, options = get_model("example1");

julia> runsim(sim, verbose=SILENT);

julia> tstep = get_well_rates(sim, "P1", "Time");

julia> qo = get_well_rates(sim, "P1", "ORAT");

julia> length(tstep) > 0
true
julia> length(qo) > 0
true

ResSimAD.SimMaster.get_state_map — Function

get_state_map(sim::Sim, var::String, t::Float64)

Get state map for variable var at time t.

Examples

julia> using ResSimAD: get_model, runsim, get_state_map, get_well_rates, SILENT

julia> sim, options = get_model("example1");

julia> runsim(sim, verbose=SILENT);

julia> t = get_well_rates(sim, "P1", "TIME");

julia> po = get_state_map(sim, "po", t[end]);

julia> size(po)
(450,)

ResSimAD.SimMaster.add_well — Function

add_well(sim::Sim, welltype::String, well_option::Dict)

Add one well of type welltype, with well setups specified in well_option

Arguments

sim::Sim: Sim object

welltype::String: type of the well (case insensitive)

"injector": injector

"producer": producer

well_option::Dict: well setups with keys

"name" -> String, name of the well

"perforation" -> Vector{Tuple{Int, Int, Int}}, indices for perforated blocks

"radius" -> Float64, well radius

"mode" -> String, well control mode

"target" -> Float64, well constrol target

Examples

julia> using ResSimAD: get_model, add_well

julia> sim, options = get_model("example1");

julia> length(keys(sim.facility))
2

julia> p2 = Dict("name" => "P2", "perforation"=>[(5,5,1)], "radius"=>0.5, "mode"=>"bhp", "target"=>5600.);

julia> add_well(sim, "producer", p2);

julia> length(keys(sim.facility))
3

ResSimAD.SimMaster.shut_well — Function

shut_well(sim::Sim, name::String)

Shut well name

ResSimAD.SimMaster.change_dt — Function

change_dt(sim::Sim, dt::Float64)

Set next time step size to be dt